move Echarts loader to its own loader file
